Projet QCM
This is my first java program for studies. I had to create a program with out POO. Multiple Choice Questions with one or several good answers. the user had to fill the questions, possible answers, good answers. Then the user take the place of the student and answer all the questions he filled. After that he has a notification if he pass or not.

PS:All the commentary and String msg are in French.